refactor(recruiting): split application table components
